Move time stringifiers to sub-package.
Also made utility classes `final`.
