+++ client/src/main/java/org/spine3/protobuf/PackingIterator.java
@@ -32,7 +32,7 @@ import java.util.Iterator;
  * @author Alexander Yevsyukov
  * @see AnyPacker#pack(Iterator)
  */
-class PackingIterator extends UnmodifiableIterator<Any> {
+final class PackingIterator extends UnmodifiableIterator<Any> {
 
     private final Iterator<Message> source;
