+++ client/src/main/java/org/spine3/time/Formats.java
@@ -33,7 +33,7 @@ import static org.spine3.time.Timestamps2.NANOS_PER_MILLISECOND;
  *
  * @author Alexander Yevsyukov
  */
-class Formats {
+final class Formats {
 
     static final String DATE_FORMAT = "yyyy-MM-dd";
     static final String DATE_TIME_FORMAT = "yyyy-MM-dd'T'HH:mm:ss";
