+++ client/src/test/java/org/spine3/string/DurationStringifierShould.java
@@ -22,6 +22,7 @@ package org.spine3.string;
 
 import com.google.protobuf.Duration;
 import org.junit.Test;
+import org.spine3.string.time.TimeStringifiers;
 import org.spine3.time.Durations2;
 
 import static org.junit.Assert.assertEquals;
