+++ client/src/test/java/org/spine3/string/ZoneOffsetStringifierShould.java
@@ -21,6 +21,7 @@
 package org.spine3.string;
 
 import org.junit.Test;
+import org.spine3.string.time.TimeStringifiers;
 import org.spine3.time.ZoneOffset;
 
 import static org.junit.Assert.assertEquals;
